my-project/User-Service/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pymysql==1.1.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0

# 日志
python-json-logger==2.0.7

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 文件存储
boto3==1.34.0
minio==7.2.0 